backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0

# Image processing
Pillow==10.1.0

# AI/ML
openai>=1.0

# Other utilities
aiofiles==23.2.1
pydantic==2.5.0
